[lib] Update `messageStore.messages` with `msgStoreOps` for [logOut/deleteAcc/leaveThr/etc].success actions
Summary:
Instead of getting `messageStore.messages` from the `filteredMessageStore` returned by `updateMessageStoreWithLatestThreadInfos(...)`... we use the `messageStoreOps` returned by `updateMessageStoreWithLatestThreadInfos(...)` to compute the new message store.

Note that a `messageStore.messages` is still being created within `updateMessageStoreWithLatestThreadInfos(...)` and will be removed in a subsequent diff

---

Depends on D4201

Test Plan: This should be a straightforward refactor given we've had `assertMessageStoreMessagesAreEqual` in place for some time now.

Reviewers: def-au1t, palys-swm

Subscribers: ashoat, Adrian, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4202