[lib] Update `messageStore.messages` with `msgStoreOps` for processServerRequests action
Summary:
Instead of getting `messageStore.messages` from the `messageStoreAfterReassignment` returned by `reassignMessagesToRealizedThreads(...)`... we use the `messageStoreOps` returned by `reassignMessagesToRealizedThreads(...)` to compute the new `messageStore.messages`.

Note that a `messageStore.messages` is still being created within `reassignMessagesToRealizedThreads(...)` and will be removed in a subsequent diff

---

Depends on D4202

Test Plan: This should be a straightforward refactor given we've had `assertMessageStoreMessagesAreEqual` in place for some time now.

Reviewers: def-au1t, palys-swm

Subscribers: ashoat, Adrian, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4203