[lib] Update `messageStore.messages` with `msgStoreOps` for createLocalMessage
Summary:
Instead of using `messageStore.messages` from the previous `updatedMessageStore` object... we pull it from the `processedMessageStore` object.

We actually remove `updatedMessageStore` altogether and include the changes to `updatedMessageStore.threads` in the `messageStore` of the returned object.

---

Depends on D4203

Test Plan:
This should be a straightforward refactor given we've had `assertMessageStoreMessagesAreEqual` in place for some time now.

Reviewers:
def-au1t, palys-swm