[services] Tunnelbroker - Fix arguments table format for new amqp queue
Summary:
This diff fixes the following bug:

When the rabbitMQ queue for the `tunnelbroker-id` is not created (clean rabbitMQ instance) there is an error when creating a new one:

```
AMQP: Queue creation error: PRECONDITION_FAILED - inequivalent arg 'x-expires' for queue 'tunnelbroker1' in vhost '/'.
```

We must remove conversion to the `string` when initializing the queue attribute. That fixes the error.

Test Plan:
1. On the master branch where rabbitMQ Docker container instance was not initialized before (was clean) run `yarn run-tunnelbroker-service-dev-mode` and get a queue creation error (which is described above in the diff description).

2. Apply the diff patch and re-run `yarn run-tunnelbroker-service-dev-mode`. The rabbitMQ queue will be created successfully.

Reviewers: karol-bisztyga, palys-swm

Reviewed By: palys-swm

Subscribers: ashoat, palys-swm, Adrian, atul, karol-bisztyga, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4168