[keyserver] Create dc.sh script
Summary:
We're going to want Docker hosts to run this script instead of `docker compose` and `docker-compose` commands. More details on why it's necessary in DSOMETHING and [ENG-1230](https://linear.app/comm/issue/ENG-1230/keyserver-mysql-backups-not-saving-on-linux-docker-hosts).

Making this a bash script instead of a `yarn` script so that Docker hosts don't need to install yarn in order to run it.

Depends on D4212

Test Plan: I tested this in concert with D4211

Reviewers: palys-swm, atul

Subscribers: Adrian, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4213