[web] Create AddUsersList component that fetches users from the store
Summary:
This component is responsible for displaying a list of users that can be added to friend / block list. To get a list of users we start by searching our store.

Depends on D4266

Test Plan: Render the component and check if all the users that match search text were returned.

Reviewers: def-au1t, atul, benschac

Subscribers: ashoat, Adrian, yayabosh

Differential Revision: https://phab.comm.dev/D4267