[web] Add option to filter users by status
Summary:
We don't want to display users that are friends when adding new friends, just like we don't want to do it for blocked users. To allow that, we can specify a set of excluded statuses.

Depends on D4305

Test Plan: Check if setting excluded statuses blocks the users with these from being displayed.

Reviewers: def-au1t, atul

Subscribers: ashoat, Adrian, yayabosh

Differential Revision: https://phab.comm.dev/D4306