[chore] Change "M1" references to "ARM-based Mac(s)"
Summary:
Minor change -- changed all occurrences of "M1" (not in code, in
comments and documentation) to "ARM-based Mac(s)". This is because now
that M2 exists, M1 only classifies the the first generation of Apple
silicon Macs.

Test Plan:
N/A, these were all comments and documentation so these changes
are all only cosmetic.

Reviewers: atul

Subscribers: ashoat, palys-swm, Adrian

Differential Revision: https://phab.comm.dev/D4244