[web] Add buttons that open new modals
Summary: Add label and callback props to user list modal so that new modals can be opened. Set these props in block / friend list to appropriate modals.

Depends on D4313

Test Plan: Open friend list and check if it has a correct button that opens a modal where new friends can be added.
Open block list and check if it has a correct button that opens a modal where users can be blocked.

Reviewers: def-au1t, atul