[keyserver] Extract getBackupInfos
Summary:
We're going to want to reuse some of this code for [ENG-1234](https://linear.app/comm/issue/ENG-1234/add-functionality-to-limit-disk-space-usage-of-keyserver-mysql-backups), so I extracted it.

Depends on D4262

Test Plan: I copy pasted the new `backups.js` file into prod and added a `console.log` of the result of `getBackupInfos`. I made sure it was in chrono order and had everything right

Reviewers: atul, yayabosh

Subscribers: palys-swm, Adrian

Differential Revision: https://phab.comm.dev/D4263