[services] Tests - Initialize project
Summary:
Please note that this is a draft, I'm checking for your opinions about the idea.

This initializes the project:
- adds dependencies - you'll see me using them later in the stack
- adds the port for the test "service"
- adds the build script - the idea is that we're going to pass the target service as an argument to the test "service" and in the testing container we're going to compile files only for the proto of the target service.

I think we can treat the test project as another service that will never leave the dev environment. This is because we do not want devs to download dependencies and build the project on their host machines, it's going to be built inside of the docker container.

Test Plan: This change itself doesn't do anything, so for now, none.

Reviewers: varun, palys-swm, geekbrother

Reviewed By: palys-swm

Subscribers: geekbrother, ashoat, palys-swm, Adrian, atul, yayabosh

Differential Revision: https://phab.comm.dev/D4161