[services] Backup - Stop suppressing invalid chunk structure
Summary:
Depends on D4228

We should not break the reactor connection if we store a chunk in the database. In this case, we never catch the error when the client tries to push two consecutive chunks below the database limit and that should be pointed out and returned back to the client. If we suppress it, the client may never know that the second chunk is lost.

Test Plan: on the client, send two log chunks below the database limit which is 400kb.

Reviewers: palys-swm

Subscribers: ashoat, palys-swm, Adrian, atul, yayabosh

Differential Revision: https://phab.comm.dev/D4232