[services] Backup - Add current id and attachment holders
Summary:
Depends on D4233

We need to be able to separate logs from each other when we pull the backup from the cloud.
For now, we just read the following log chunks without knowing how many logs we really get since multiple chunks can fit into one log.

With the current id, we can keep track of what is the current log of the chunk that is being pushed so that when it changes, we know that the current chunk goes for a different log.

Test Plan: none - it is going to be used later in the stack. Services still can be built.

Reviewers: palys-swm, ashoat

Subscribers: ashoat, palys-swm, Adrian, atul, yayabosh

Differential Revision: https://phab.comm.dev/D4245