[docs] Replace `wget` command in gRPC docs with `curl`
Summary: Linear issue with full context here. The current gRPC docs make
use of the `wget` command to install protobuf from GitHub, but `wget` is
not pre-installed on macOS. Replaced `wget` with `curl -L`.

Test Plan: Re-ran all the commands before and after the `wget` command in
the `comm` home directory with the new `curl` command, made sure outputs
were the same as with the
`wget` command.

Reviewers: ashoat!, varun!

Subscribers: