[web] Fix sidebar title cutoff in `ChatThreadList`
Summary: Seems to be a regression, not sure when it was introduced exactly. This diff repositions the `ChatThreadListItemMenu` `.menuSidebar` `div` to be absolutely positioned so that it doesn't interfere with the `ChatThreadListSidebar`. What was happening before was the `ChatThreadListSidebar` was being covered by the `ChatThreadListItemMenu`. It probably went unnoticed since it wasn't tested with a sidebar message that was long enough to trigger the ellipsis overflow.

Before:

After:

Test Plan: Looks as expected on Chrome/Safari.

Reviewers: atul, ashoat

Subscribers: