[web] Add `LoadingIndicator` to `ThreadSettingsGeneralTab` Save button
Summary: Related Linear issue here. Adds a `LoadingIndicator` to the Save button in `ThreadSettingsGeneralTab` to indicate that a change is in progress. Depends on D4105 because of the renaming of `inputDisabled` to `threadSettingsOperationInProgress`.

Before:

After:

Test Plan: Tested on Chrome/Safari and works as expected. However, since `localhost/comm` is faster than `prod`, the change can be seen less quickly in the videos. However, it is very similar to D3605.

Reviewers: atul, ashoat

Subscribers: