[keyserver] Add `sendmail` installation to `Dockerfile`
Summary:
The `SubscriptionForm` on the landing page is broken and is displaying "Request failed" when an email is entered.

The keyserver Dockerfile is based on `node:16.13-bullseye` which is based on Debian bullseye which doesn't come with `sendmail` installed.

While we may still want to move from `sendmail` to Postmark, this might be a good intermediate solution to get things working again... assuming it works.

Test Plan:
1. Get landing page working on Debian VM
2. Ensure that entering an email into the `SubscriptionForm` fails on submission
3. Install `sendmail`
4. See if entering an email into the `SubscriptionForm` and submitting succeeds

(this is in progress)

Reviewers:
ashoat, def-au1t, palys-swm