[lib] Remove extraneous ternary from `pluralize` function in `text-utils`
Summary: The ternary condition in the `pluralize` function in `text-utils` is unnecessary since its "else" condition is unreachable. Because of the order of the cases in `pluralize`, if the `maxNumberOfNouns === 1` conditional is entered, `nouns.length` will always be greater than 1, since the prior two conditionals check if `nouns.length` equals 0 or 1. Therefore, it can be removed.

Test Plan: N/A, but I made sure to note that the `length` property of an array can never be a negative number. So `nouns.length` is guaranteed to be greater than 1 if the `maxNumberOfNouns === 1` conditional is reached.

Reviewers: atul, palys-swm

Subscribers: