[lib] Update `messageStore.messages` with `msgStoreOps` for remaining action types
Summary:
These are the last actions in `reduceMessageStore` that need to be modified so we can remove `assertMessageStoreMessagesAreEqual` and solely use the "ops approach" to update `messageStore.messages` in redux.

The two remaining instances of `assertMessageStoreMessagesAreEqual` are in `mergeNewMessages` and will be addressed in the next diff.

---

D4330

Test Plan:
This should be a straightforward refactor given we've had `assertMessageStoreMessagesAreEqual` in place for some time now.

Reviewers:
def-au1t, palys-swm, yayabosh