[lib] Replace `mergedStore.messages` with `processedStore.messages` in `mergeNewMessages`
Summary:
This diff removes the second-to-last usage of `assertMessageStoreMessagesAreEqual(...)` in `message-reducer`. The last usage is also in `mergeNewMessages`, but the logic there is more involved so I'm going to take a second to read through it again and put up a seperate diff.

---

Depends on D4331

Test Plan: This should be a straightforward refactor given we've had `assertMessageStoreMessagesAreEqual` in place for some time now.

Reviewers: def-au1t, palys-swm, yayabosh

Subscribers: ashoat, Adrian

Differential Revision: https://phab.comm.dev/D4332