[lib] Remove `assertMessageStoreMessagesAreEqual(...)` from `message-reducer`
Summary:
We no longer need to assert that the "old approach" and the "ops approach" are equal... because we're removing the "old approach" altogether.

This diff
- removes the last usage of `assertMessageStoreMessagesAreEqual` from `mergeNewMessages`
- removes `assertMessageStoreMessagesAreEqual`
- removes imports that `assertMessageStoreMessagesAreEqual` depended on
- removes `actionType` argument from `mergeNewMessages(..._)` since it was only used by `assertMessageStoreMessagesAreEqual` from within `mergeNewMessages`

---

Depends on D4332

Test Plan:
This should be a straightforward refactor given we've had `assertMessageStoreMessagesAreEqual` in place for some time now.

Reviewers:
def-au1t, palys-swm, yayabosh