[lib] Rename `threadsToMessageIDsFromMessageInfos` to `mapThreadsToMessageIDsFromOrderedMessageInfos`
Summary:
I think this is a better, more descriptive name. I think it's good to include the fact that the `messageInfos` are ordered in the name so it's more evident at the callsite.

This is an "opinionated" change, so open to abandoning if others disagree.

Test Plan:
flow

Reviewers:
def-au1t, palys-swm, yayabosh