[native] [web] Improve wording when clearing thread name or description
Summary: Related Linear issue with full context here. Improved the wording of the `robotext` that is outputted when setting the thread name or description to the empty string.

Test Plan:
Looks as expected on `web` and `native`, also changed thread name and description on `web` and `native` and both platforms match/work as expected:

Also tested to make sure the notification text matches the updated `robotext` (thanks @palys-swm for your help here!):

Reviewers: atul, palys-swm

Subscribers: ashoat, Adrian, palys-swm

Differential Revision: https://phab.comm.dev/D4346