[web] Make button size consistent between variants
Summary:
Our button has different size depending on the variant, because secondary button has a border that increases its size. The issue isn't usually noticeable, but can be detected by using the debugger. It becomes visible when secondary and primary buttons are displayed in the same line.
The solution is for button to always have a border, which is transparent when the button is primary.

Test Plan: Checked in a couple of places if it still looks ok. There are places when sizing is slightly changed, because primary button is 2 pixels bigger, but the layout is ok in every place.

Reviewers: def-au1t, atul, yayabosh

Subscribers: ashoat, Adrian

Differential Revision: https://phab.comm.dev/D4353