[web] Remove `safari-pinned-tab.svg`
Summary:
Addresses https://linear.app/comm/issue/ENG-963/remove-old-squad-cal-favicon-on-safari

It used to be that Safari didn't display favicons for each tab and instead showed the page title. However, if a tab was pinned--or there wasn't enough space to show an abbreviated title--it would display the first letter of the page title:

(examples)

Apple gave the option of providing a black square SVG (`safari-pinned-tab.svg`) to be displayed for "pinned" tabs. There was also the option to specify a color for the SVG to be tinted when the tab was active.

However, recent versions of Safari display favicons for tabs across the board (it's no longer even possible to disable them), and including `safari-pinned-tab.svg` is no longer necessary... but Safari will prioritize it if it's provided. Based on what I've read, it seems like the only reason to specify `safari-pinned-tab.svg` is if you want something other than the default favicons to be displayed on Safari. I don't think that's something we care about.

Test Plan: Ensured correct favicon displayed in Safari/Chrome after clearing cache.

Reviewers: yayabosh, palys-swm, def-au1t

Subscribers: ashoat, Adrian

Differential Revision: https://phab.comm.dev/D4358