[services][identity] register_user RPC
Summary: Implementation of the register_user RPC. We only need to do this for PAKE -- there is no registration process for wallet auth. The bulk of the logic is performed in the subroutines. This function handles updating some state and sending messages to the outbound stream.

Test Plan: called the RPC and tested all failure cases as well as the success case. subsequently tried logging in and received a new access token

Reviewers: jimpo!, palys-swm!, karol-bisztyga

Subscribers: