[ignore] Try relative path instead of absolute for `native/.flowconfig`
Summary:
Right now the `.flowconfig` for the `native` module expects the root directory to be named `comm`. This is has been an issue with Buildkite in the past because it name sthe root directory after the name of the "pipeline."

We previously got around this by naming the flow "pipeline" to "comm".

However, now that we're using the Buildkite docker plugin, it names the root directory "workdir" which causes problems with `flow`.

This diff is to test whether changing the path from absolute to relative fixes things.

(It's in the draft state because I'm just experimenting and haven't fully read up on the glob rules to ensure that the before/after are fully equivalent).

Test Plan:
CI

Reviewers:
palys-swm, def-au1t, yayabosh