[services] Tunnelbroker - Add test for putMessageItemsByBatch function
Summary:

This is a test for `putMessageItemsByBatch` function introduced in diff.
In this test we are creating a messages batch of 29 items and put them into the database by the `putMessageItemsByBatch` function. We are expecting the
size of the found items would be the same as a items batch size.

Test Plan:

Run `yarn test-tunnelbroker-service-dev-mode`

Reviewers:

Subscribers: