[docs] Remove references to create-db.js script
Summary: Following D4151, the keyserver migration code should automatically setup the database when it sees the `meta` table is missing. As a result, it's no longer necessary for the user to manually call `create-db.js`.

Test Plan: As part of D4151 I tested that the database gets automatically set up if it has no tables

Reviewers: varun, atul, palys-swm

Subscribers: