[CI] Add `identity_build` workflow to Buildkite/GH Actions
Summary:
This was previously handled as part of the "Services CI" workflow that will be deprecated later in this stack. This diff allows us to increase concurremcy as we move over to the autoscaling EC2 instances.

Differences on the Buildkite side:
1. We use the actual shell command instead of `yarn blah` because the autoscaling instances are intended to be as minimal as possible an don't include `node`, `yarn`, etc.

2. We use the `--no-cache` flag to force a new complete build on each run. The previous "Services CI" workflow depended heavily on caches, because running 3-4 complete builds would take a long time. Now, we can separate these out and take advantage of the concurrency provided by the autoscaling instances.

---

Depends on D4390

Test Plan:
Buildkite: Copy/pasted from what I have in the Buildkite GUI that has been "active" for the last 2-3 days without issue.

GitHub: Will see what happens when it lands, but fairly confident it should be good to go.

Reviewers:
def-au1t, palys-swm, yayabosh