[services] Lib - fix bugs in ClientWriteReactorBase
Summary:
Depends on D4252

I spotted some obvious bugs in this class. They've not been spotted before because the compilation doesn't see them. I guess it's because we do not yet use this class.

Test Plan: None really. It doesn't affect the compilation, but I wanted to make these two right so if we ever use this class, it will be better.

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: ashoat, palys-swm, Adrian, atul, yayabosh

Differential Revision: https://phab.comm.dev/D4270