[services] Add teardown script
Summary:
Depends on D4271

Sometimes we may want to quickly delete all the containers and data connected to the local cloud. This may come useful when we do some tests that perform the following:
- add
- get
- remove

and the test fails on `get`. The `remove` will never be performed, so next time we run this test, `add` may fail.

That's just one example, I decided this may be useful in other situations as well.

Test Plan:
```
cd services
yarn run-local-cloud
```
then do some operations on the local cloud and see the records in the dynamo db or objects on the s3
```
yarn tear-down-local-cloud
yarn run-local-cloud
```
see that the objects/records are gone

Reviewers: palys-swm, varun, ashoat

Reviewed By: palys-swm, ashoat

Subscribers: ashoat, palys-swm, Adrian, atul, yayabosh

Differential Revision: https://phab.comm.dev/D4272