[native] Increase clickable region of `ColorSelector` Save button
Summary: Quick fix before finalizing `ColorSelector`. Noticed a bug where the "Save" button in the `ColorSelector` was only being clicked when clicking on the "Save" text, and not anywhere on the button.

Fixed this by removing the outer `View` container and replacing it with a `TouchableOpacity`, allowing the entire region to be clickable **and** execute the `onPress` function.

What it looks like now:

Test Plan: Looks and works as expected on iOS.

Reviewers: atul, palys-swm

Subscribers: