[web] Rename `MessageActionButtons` to `MessageTooltip`
Summary: Relevant Linear issue [[ https://linear.app/comm/issue/ENG-1338/rename-messageactionbuttons-to-messagetooltip | here ]]. @ashoat and I discussed this today and determined that since the timestamp was moved into `MessageActionButtons` (see D3571) it doesn't make sense that the component called `MessageActionButtons` contains the action buttons **and** the timestamp. Instead, renamed it to something that better encapsulates the component, which is a tooltip that opens when a user hovers over a message or `RobotextMessage`.

Test Plan: Replaced all occurrences of `MessageActionButtons`, `messageAction` that had to do with `MessageActionButtons`. Did not replace the instances of `messageActionButtons` in the CSS for `MessageActionButtons` because those //were// actually pertaining to the message ation buttons. Everything else (like in `ComposedMessage` or `RobotextMessage`) were just referring to the tooltip as a whole.

Reviewers: atul, palys-swm

Subscribers: