[keyserver] Run `source-nvm.sh` using `bash` in `Docker` container
Summary: Relevant Linear issue with full context [[ https://linear.app/comm/issue/ENG-1339/run-source-nvmsh-using-bash-in-docker-container | here ]]. This diff modifies the `Dockerfile` in `keyserver` to execute `source-nvm.sh` in the environment specified by the shebang in `source-nvm.sh` instead of in the current environment of the container, which uses `sh`.

Test Plan: Ran the Docker keyserver build locally and it worked with no errors.

Reviewers: atul, ashoat

Subscribers: