[server] Support pending threads when calculating `finalNavInfo`
Summary:
Currently, on server side, when calculating initial navInfo, there is a check that redirects user to latest thread, if the user has no permission to read the thread.
However, there is no check if thread with ID calculated from URL even exist in store. As we want to support navigation with pending threads, the non-existing thread IDs should not be removed on server side.

Test Plan: The app behavior remains unchanged, if non-existing ID is entered, the user is still redirected correctly. Support for pending threads on React side will be introduced in following diffs.

Reviewers: palys-swm, atul

Subscribers: ashoat, Adrian, yayabosh

Differential Revision: https://phab.comm.dev/D4435