Retrieve messages from temporal message storage in AppDelegate
Summary: This differential uses TemporalMessageStorage API to retrieve messages and store them in SQLite before JavaScript is started.

Test Plan: Send several messages when mobile app is not running. Launch app in a debugger and ensure that all messages are retrieved from temporal message store. Then before last message is written to the database, download database from XCode and ensure that all messages seen so far are written to the database.

Reviewers: palys-swm, karol-bisztyga, varun

Subscribers: