[services] replace rusoto with official aws rust sdk
Summary: swap out the rusoto dynamodb client with the official aws one. we need to do this because rusoto is no longer actively maintained. this diff is lengthy, but there are no changes to the actual code logic.

Test Plan: cargo build, cargo run, call the RPCs and check that the DDB tables have the expected data

Reviewers: palys-swm!, karol-bisztyga, jimpo, ashoat!

Subscribers: Adrian, atul, yayabosh

Differential Revision: https://phab.comm.dev/D4459