[services] wrap database client in Arc to avoid creating a new client each time we clone
Summary: According to this [discussion](https://github.com/awslabs/aws-sdk-rust/discussions/564), the SDK clients are intended to be reused. We should create one client in main and then wrap it in an Arc, so that we reuse the underlying client struct when we clone.

Test Plan: Confirmed that I could make requests to DDB with the wrapped client

Reviewers: jimpo, karol-bisztyga, palys-swm!

Subscribers: