[web] Introduce onMessageSent callbacks
Summary:
Introduce ability to register/unregister callbacks that are executed when user sends a message. This functionality exists in mobile app already.
The feature is required to detect sending message and change chat creation view to chat view automatically after sending message.

The code from `input-state-container` on web seems to be very similar to the native one (probably it could be copy-pasted at some point in the past).
Maybe it would be a good idea to refactor it so the platforms would share common code related to input state. However  I think it is out of the scope of this diff.
Test Plan:
Tested the callbacks after introducing chat creation view. It's not used yet.
Reviewers:
palys-swm, atul
Subscribers: