[services] Tunnelbroker - Switching to use composite key in a messageItem entity
Summary:

To avoid possible message UUID collision and speedup batch messages deletion we should switch to using the composite key for the `tunnelbroker-messages`
table which will consist of `messagesID` as a Sort Key and `toDeviceID` as a Partition key.

This diff is a part of the stack and introduces changes to `MessageItem` database entity.

Linear task: [[ https://linear.app/comm/issue/ENG-1362/update-to-use-composite-key-messageidtodeviceid-for-messages-table-in | ENG-1362 ]]

Test Plan:

Will be populated soon...

Reviewers:

Subscribers: